pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "exiora-ai-chatbot"
version = "2.0.0"
description = "Next Generation Multimodal AI Chatbot with Text, Image, and Voice Capabilities"
authors = [
    {name = "Saad Sadik Shaikh", email = "your-email@example.com"}
]
readme = "README.md"
requires-python = ">=3.8,<3.12"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "streamlit>=1.24.0,<2.0.0",
    "openai>=1.3.0,<2.0.0",
    "google-generativeai>=0.3.0,<1.0.0",
    "groq>=0.4.0,<1.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
audio = [
    "gTTS>=2.3.0",
    "pyttsx3>=2.90",
    "SpeechRecognition>=3.10.0",
    "pydub>=0.25.1",
]
vision = [
    "opencv-python-headless>=4.8.0",
]
full = [
    "streamlit-option-menu>=0.3.6",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "jsonschema>=4.17.0",
    "loguru>=0.7.0",
    "cachetools>=5.3.0",
    "cryptography>=41.0.0",
    "typing-extensions>=4.7.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["utils*"]
